(#36) Refactor commands_* files
